refactor indestructible registry
